don't report connection still alive after close without Client.disconnect
previously, due to an oversight, closing the connection through any means
other than calling the function Client.disconnect (i.e. computer
suspending, server dying) would cause AP.js's state machine to still report
that it was connected to the server. This fixes that by moving much of
the deinitialization logic in Client.disconnect to a new disconnection
listener which is established in Client.#finalizeConnection (so that it
only gets called for fully initialized socket data).